Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -35,7 +35,8 @@ class QuestionsController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
+     *
      * @return \Illuminate\Http\Response
      */
     public function store(AskQuestionRequest $request)
@@ -48,7 +49,8 @@ class QuestionsController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Question  $question
+     * @param  \App\Question $question
+     *
      * @return \Illuminate\Http\Response
      */
     public function show(Question $question)
@@ -61,23 +63,32 @@ class QuestionsController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Question  $question
+     * @param  \App\Question $question
+     *
      * @return \Illuminate\Http\Response
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            return abort(403, "Access denied");
+        }
         return view("questions.edit", compact('question'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Question  $question
+     * @param  \Illuminate\Http\Request $request
+     * @param  \App\Question            $question
+     *
      * @return \Illuminate\Http\Response
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            return abort(403, "Access denied");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', "Your question has been updated.");
@@ -93,6 +104,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            return abort(403, "Access denied");
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', "Your question has been deleted.");
